requirements.txt
fastapi
fastapi_limiter
slowapi
fastapi[security]
uvicorn
sqlalchemy
pydantic
alembic
pandas
pinecone-client[grpc]
aioredis

# Logging
loguru

# Security Library
python-jose

# Langchain Library
langchain
langchainhub
langchain_core
langchain_community
langchain-google-genai # This package contains the LangChain integrations for the Gemini API.
Pillow # PIL image-loading library
groq
tiktoken
chromadb

# Scrapping Library
beautifulsoup4

#Tools
PyMuPDF